+++ src/dataProvider/SparqlDataProvider.js
@@ -53,11 +53,9 @@ export default {
       });
     }
 
-    const totalItems = await query.totalItems;
-
     return {
       data: results,
-      total: parseInt(totalItems),
+      total: query.totalItems
     };
   },
   getOne: async function getOne() {
@@ -118,10 +116,14 @@ async function fetchQuery(query) {
     if (!query.variableOntology) {
       query.variableOntology = findPredicates(parsedQuery);
     }
-    if (!parsedQuery.limit) {
+    if (parsedQuery.limit !== undefined && query.offset + query.limit > parsedQuery.limit) {
+      parsedQuery.limit = parsedQuery.limit - query.offset;
+    } else {
       parsedQuery.limit = query.limit;
     }
-    if (!parsedQuery.offset) {
+    if (parsedQuery.offset) {
+      parsedQuery.offset += query.offset;
+    } else {
       parsedQuery.offset = query.offset;
     }
     if (!parsedQuery.order && query.sort && query.sort.field !== "id") {
@@ -279,12 +281,7 @@ async function handleQueryExecution(execution, query) {
     const resultType = execution.resultType;
     if (execution.resultType !== "boolean") {
       const metadata = await execution.metadata();
-      const totalItems = metadata.totalItems;
-      if (!totalItems) {
-        query.totalItems = countQueryResults(query);
-      } else {
-        query.totalItems = totalItems;
-      }
+      query.totalItems = await countQueryResults(query);
       variables = metadata.variables.map((val) => {
         return val.value;
       });
@@ -298,19 +295,27 @@ async function handleQueryExecution(execution, query) {
 /**
  *
  * @param {object} query - the query which is to be executed and additional information about the query.
- * @returns {Array<Term>} the results of the query
+ * @returns {number} the actual number of results in the query, if it were executed
  */
 async function countQueryResults(query) {
   const parser = new Parser();
   const parsedQuery = parser.parse(query.rawText);
+  const distinctInitial = parsedQuery.distinct;
+  const offsetInitial = parsedQuery.offset;
+  const limitInitial = parsedQuery.limit;
   parsedQuery.queryType = "SELECT";
+  parsedQuery.distinct = false;
+  parsedQuery.offset = 0;
+  if (parsedQuery.limit) {
+    delete parsedQuery.limit;
+  }
   parsedQuery.variables = [
     {
       expression: {
         type: "aggregate",
         aggregation: "count",
         expression: { termType: "Wildcard", value: "*" },
-        distinct: false,
+        distinct: distinctInitial
       },
       variable: { termType: "Variable", value: "totalItems" },
     },
@@ -322,7 +327,14 @@ async function countQueryResults(query) {
     fetch: fetch,
     httpProxyHandler: proxyHandler,
   });
-  return (await bindings.toArray())[0].get("totalItems").value;
+  let totalItems = parseInt((await bindings.toArray())[0].get("totalItems").value);
+  if (offsetInitial) {
+    totalItems -= offsetInitial;
+  }
+  if (limitInitial && totalItems > limitInitial) {
+    totalItems = limitInitial;
+  }
+  return totalItems;
 }
 
 const queryTypeHandlers = {